# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /composer/generate:
    post:
      operationId: generateComposition
      tags: [generate]
      summary: Generate a full AI composition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateCompositionBody"
      responses:
        "200":
          description: Composition generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateCompositionResponse"
        "503":
          description: AI service unavailable (API key not configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /composer/regenerate:
    post:
      operationId: regenerateComposition
      tags: [generate]
      summary: Regenerate a composition with optional component locks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegenerateCompositionBody"
      responses:
        "200":
          description: Composition regenerated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateCompositionResponse"
        "404":
          description: Composition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /composer/compositions/{id}:
    get:
      operationId: getComposition
      tags: [generate]
      summary: Get a composition by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Composition found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Composition"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Chord:
      type: object
      required: [position, symbol, romanNumeral, duration, midiNotes]
      properties:
        position:
          type: integer
        symbol:
          type: string
        romanNumeral:
          type: string
        duration:
          type: number
        midiNotes:
          type: array
          items:
            type: integer
    HarmonyResult:
      type: object
      required: [key, bpm, timeSignature, chords]
      properties:
        key:
          type: string
        bpm:
          type: integer
        timeSignature:
          type: string
        chords:
          type: array
          items:
            $ref: "#/components/schemas/Chord"
    MelodyNote:
      type: object
      required: [pitch, startBeat, duration, velocity]
      properties:
        pitch:
          type: integer
        startBeat:
          type: number
        duration:
          type: number
        velocity:
          type: integer
    MelodyResult:
      type: object
      required: [type, phraseCount, notes]
      properties:
        type:
          type: string
        phraseCount:
          type: integer
        notes:
          type: array
          items:
            $ref: "#/components/schemas/MelodyNote"
    DrumTracks:
      type: object
      required: [kick, snare, hihat]
      properties:
        kick:
          type: array
          items:
            type: boolean
        snare:
          type: array
          items:
            type: boolean
        hihat:
          type: array
          items:
            type: boolean
        openHH:
          type: array
          items:
            type: boolean
        clap:
          type: array
          items:
            type: boolean
    DrumPattern:
      type: object
      required: [bpm, bars, steps, tracks]
      properties:
        bpm:
          type: integer
        bars:
          type: integer
        steps:
          type: integer
        swing:
          type: number
        tracks:
          $ref: "#/components/schemas/DrumTracks"
    Composition:
      type: object
      required: [id, key, bpm, timeSignature, genre, mood, prompt, harmony, melodies, drums, styleMatch, lowConfidence, createdAt]
      properties:
        id:
          type: string
        key:
          type: string
        bpm:
          type: integer
        timeSignature:
          type: string
        genre:
          type: string
        mood:
          type: string
        prompt:
          type: string
        harmony:
          $ref: "#/components/schemas/HarmonyResult"
        melodies:
          type: array
          items:
            $ref: "#/components/schemas/MelodyResult"
        drums:
          $ref: "#/components/schemas/DrumPattern"
        styleMatch:
          type: number
        lowConfidence:
          type: boolean
        createdAt:
          type: string
    GenerateCompositionBody:
      type: object
      required: [prompt, genre, mood]
      properties:
        prompt:
          type: string
          minLength: 3
          maxLength: 500
        genre:
          type: string
        mood:
          type: string
        bpm:
          type: integer
          minimum: 60
          maximum: 200
        key:
          type: string
        timeSignature:
          type: string
    RegenerateCompositionBody:
      type: object
      required: [compositionId]
      properties:
        compositionId:
          type: string
        keepHarmony:
          type: boolean
        keepMelody:
          type: boolean
        keepRhythm:
          type: boolean
    GenerateCompositionResponse:
      type: object
      required: [compositionId, composition, lowConfidence, generationMs]
      properties:
        compositionId:
          type: string
        composition:
          $ref: "#/components/schemas/Composition"
        lowConfidence:
          type: boolean
        generationMs:
          type: integer
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string