# lib/api-zod/src/generated/types/chord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Chord {
  position: number;
  symbol: string;
  romanNumeral: string;
  duration: number;
  midiNotes: number[];
}