# lib/api-zod/src/generated/types/drumPattern.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DrumTracks } from './drumTracks';

export interface DrumPattern {
  bpm: number;
  bars: number;
  steps: number;
  swing?: number;
  tracks: DrumTracks;
}