# lib/api-zod/src/generated/types/drumTracks.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DrumTracks {
  kick: boolean[];
  snare: boolean[];
  hihat: boolean[];
  openHH?: boolean[];
  clap?: boolean[];
}