# lib/api-zod/src/generated/types/generateCompositionBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateCompositionBody {
  /**
     * @minLength 3
     * @maxLength 500
     */
  prompt: string;
  genre: string;
  mood: string;
  /**
     * @minimum 60
     * @maximum 200
     */
  bpm?: number;
  key?: string;
  timeSignature?: string;
}