# lib/api-zod/src/generated/types/generateCompositionResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Composition } from './composition';

export interface GenerateCompositionResponse {
  compositionId: string;
  composition: Composition;
  lowConfidence: boolean;
  generationMs: number;
}