# lib/api-zod/src/generated/types/melodyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MelodyNote } from './melodyNote';

export interface MelodyResult {
  type: string;
  phraseCount: number;
  notes: MelodyNote[];
}