# lib/api-zod/src/generated/types/regenerateCompositionBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegenerateCompositionBody {
  compositionId: string;
  keepHarmony?: boolean;
  keepMelody?: boolean;
  keepRhythm?: boolean;
}